Move VirtualMachineError to vm_errors
